ARB Creation Response Error
Authorize.net is returning an "Ok" response to the ARBCreationRequest even when it fails. It seems only certain failures still return an "Ok" response. So this bug was intermittent.

I have patched the function to check for the response code returned which if successful should always read "I00001" error codes start with an "E". This should effectively resolve this issue.

There is also a secondary check that ensures the subscriptionId is present. However this secondary check did not have an "ELSE" statement. So due to the bug above it was not failing gracefully. I have modified this secondary check to include a proper "ELSE" statement now.